+++ fractal_client/cmd/_user.py
@@ -45,6 +45,7 @@ def user_register(
     new_password: str,
     slurm_user: str | None = None,
     cache_dir: str | None = None,
+    project_dir: str | None = None,
     username: str | None = None,
     ssh_settings_json: str | None = None,
     superuser: bool = False,
@@ -63,6 +64,8 @@ def user_register(
         new_settings["slurm_user"] = slurm_user
     if cache_dir:
         new_settings["cache_dir"] = cache_dir
+    if project_dir:
+        new_settings["project_dir"] = project_dir
     if ssh_settings_json is not None:
         ssh_settings = _read_ssh_settings_json(ssh_settings_json)
         new_settings.update(ssh_settings)
@@ -128,6 +131,7 @@ def user_edit(
     new_username: str | None = None,
     new_slurm_user: str | None = None,
     new_cache_dir: str | None = None,
+    new_project_dir: str | None = None,
     new_ssh_settings_json: str | None = None,
     make_superuser: bool = False,
     remove_superuser: bool = False,
@@ -165,6 +169,8 @@ def user_edit(
         settings_update["cache_dir"] = new_cache_dir
     if new_slurm_user is not None:
         settings_update["slurm_user"] = new_slurm_user
+    if new_project_dir is not None:
+        settings_update["project_dir"] = new_project_dir
     if new_ssh_settings_json is not None:
         ssh_settings = _read_ssh_settings_json(new_ssh_settings_json)
         settings_update.update(ssh_settings)