+++ tests/test_dataset.py
@@ -3,7 +3,12 @@ import json
 from devtools import debug
 
 
-def test_create_dataset(invoke, tmp_path, new_name):
+def test_create_dataset(
+    invoke,
+    tmp_path,
+    new_name,
+    invoke_as_superuser,
+):
     """
     Test some specific branches of the post_dataset function and parser.
     """
@@ -19,7 +24,8 @@ def test_create_dataset(invoke, tmp_path, new_name):
 
     res = invoke(
         (
-            f"project add-dataset {project_id} {new_name()} /tmp "
+            f"project add-dataset {project_id} {new_name()} "
+            "--zarr-dir /tmp "
             f"--filters {file_filters}"
         )
     )
@@ -28,7 +34,15 @@ def test_create_dataset(invoke, tmp_path, new_name):
     assert res.retcode == 0
     assert res.data["filters"] == FILTERS
 
-    res = invoke(f"--batch project add-dataset {project_id} {new_name()} /tmp")
+    # Add a project_dir to user-settings
+    res = invoke("--batch user whoami")
+    assert res.retcode == 0
+    user_id = res.data
+    res = invoke_as_superuser(
+        f"user edit {user_id} --new-project-dir /something"
+    )
+    assert res.retcode == 0
+    res = invoke(f"--batch project add-dataset {project_id} {new_name()}")
     debug(res.data)
     assert res.retcode == 0
 
@@ -37,7 +51,9 @@ def test_edit_dataset(invoke, tmp_path, new_name):
     res = invoke(f"project new {new_name()}")
     project_id = res.data["id"]
 
-    res = invoke(f"project add-dataset {project_id} {new_name()} /tmp")
+    res = invoke(
+        f"project add-dataset {project_id} {new_name()} --zarr-dir /tmp"
+    )
     dataset_id = res.data["id"]
 
     NAME = new_name()
@@ -65,7 +81,9 @@ def test_delete_dataset(invoke, new_name):
     res = invoke(f"project new {new_name()}")
     project_id = res.data["id"]
 
-    res = invoke(f"project add-dataset {project_id} {new_name()} /tmp")
+    res = invoke(
+        f"project add-dataset {project_id} {new_name()} --zarr-dir /tmp"
+    )
     dataset_id = res.data["id"]
 
     # Delete dataset
@@ -80,7 +98,9 @@ def test_show_dataset(invoke, new_name):
     # Create a project with its default dataset
     res = invoke(f"project new {new_name()}")
     project_id = res.data["id"]
-    res = invoke(f"project add-dataset {project_id} {new_name()} /tmp")
+    res = invoke(
+        f"project add-dataset {project_id} {new_name()} --zarr-dir /tmp"
+    )
     dataset_id = res.data["id"]
 
     res = invoke(f"dataset show {project_id} {dataset_id}")