+++ tests/test_user.py
@@ -76,11 +76,14 @@ def test_register_with_ssh_settings(invoke_as_superuser, new_name, tmp_path):
             ),
             f,
         )
+    PROJECT_DIR = "/somewhere/"
     res = invoke_as_superuser(
         f"user register {EMAIL_USER} {PWD_USER} "
+        f"--project-dir {PROJECT_DIR} "
         f"--ssh-settings-json {valid_json}"
     )
     assert res.retcode == 0
+    assert res.data["settings"]["project_dir"] == PROJECT_DIR
     assert res.data["settings"]["ssh_host"] == "SSH_HOST"
     assert res.data["settings"]["ssh_private_key_path"] == (
         "/SSH_PRIVATE_KEY_PATH"
@@ -254,9 +257,11 @@ def test_edit_user_settings(invoke_as_superuser, tmp_path, new_name):
         "slurm_user": None,
         "slurm_accounts": [],
         "cache_dir": None,
+        "project_dir": None,
     }
     SSH_HOST = "something.somewhere"
     SSH_PRIVATE_KEY_PATH = "/tmp/something.key"
+    NEW_PROJECT_DIR = "/somewhere/else/"
     NEW_USER_SETTINGS = {
         "ssh_host": SSH_HOST,
         "ssh_username": None,
@@ -266,6 +271,7 @@ def test_edit_user_settings(invoke_as_superuser, tmp_path, new_name):
         "slurm_user": None,
         "slurm_accounts": [],
         "cache_dir": None,
+        "project_dir": NEW_PROJECT_DIR,
     }
 
     # Register a new user
@@ -296,6 +302,7 @@ def test_edit_user_settings(invoke_as_superuser, tmp_path, new_name):
         )
     cmd = (
         f"user edit {user_id} "
+        f"--new-project-dir {NEW_PROJECT_DIR} "
         f"--new-ssh-settings-json {ssh_settings_file.as_posix()}"
     )
     res = invoke_as_superuser(cmd)
@@ -353,40 +360,6 @@ def test_edit_arguments(invoke_as_superuser):
     with pytest.raises(SystemExit):
         cmd = "user edit SOME_USER_ID --make-superuser --remove-superuser"
         invoke_as_superuser(cmd)
-
-
-@pytest.mark.skip(
-    reason="Delete-user endpoint was removed in fractal-server 1.4.0"
-)
-def test_delete_as_user(invoke, invoke_as_superuser, caplog, new_name):
-    EMAIL_USER = f"{new_name()}@example.org"
-    # Register a new user
-    res = invoke_as_superuser(f"user register {EMAIL_USER} {PWD_USER}")
-    user_id = res.data["id"]
-    # Call fractal user edit
-    with pytest.raises(SystemExit):
-        invoke(f"user delete {user_id}")
-    debug(caplog.text)
-    assert "403" in caplog.text
-
-
-@pytest.mark.skip(
-    reason="Delete-user endpoint was removed in fractal-server 1.4.0"
-)
-def test_delete_as_superuser(invoke_as_superuser, caplog, new_name):
-    EMAIL_USER = f"{new_name()}@example.org"
-    # Register a new user
-    res = invoke_as_superuser(f"user register {EMAIL_USER} {PWD_USER}")
-    user_id = res.data["id"]
-    # Call fractal user delete
-    res = invoke_as_superuser(f"user delete {user_id}")
-    assert res.retcode == 0
-    # Check that user was not found
-    with pytest.raises(SystemExit):
-        res = invoke_as_superuser(f"user show {user_id}")
-    debug(caplog.text)
-    assert "404" in caplog.text
-    assert "Not Found" in caplog.text
 
 
 def test_whoami_as_user(invoke, tester):